REFACTOR summary/* + TESTS PASSED
