+++ src/lib/validatePartnerBonusForm.ts
@@ -1,6 +1,6 @@
 import { PartnerUserInput } from '../types/PageUserInputs'
 import { parseInputNumber } from './utils'
-import { PARTNER_MAX_ODPOCET } from './calculation'
+import { PARTNER_MAX_ODPOCET } from '../plugins/summary/calculation/calculation'
 
 export const validatePartnerBonusForm = (
   values: PartnerUserInput,
