+++ src/pages/_app.tsx
@@ -9,7 +9,7 @@ import '../styles/libs.css'
 import React, { useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
 import { setLocale } from 'yup'
-import { calculate } from '../lib/calculation'
+import { calculate } from '../plugins/summary/calculation/calculation'
 import { TaxFormUserInput } from '../types/TaxFormUserInput'
 import { TaxForm } from '../types/TaxForm'
 import Layout from '../plugins/main/_layout/Layout'
