+++ src/pages/deti.tsx
@@ -1,4 +1,3 @@
-import { validate, Deti } from '../plugins/steps/3b-deti/deti'
+import { Deti } from '../plugins/steps/3b-deti/deti'
 
-export { validate }
 export default Deti
