+++ src/pages/dochodok.tsx
@@ -1,4 +1,3 @@
-import { validate, Dochodok } from '../plugins/steps/4-dochodok/dochodok'
+import { Dochodok } from '../plugins/steps/4-dochodok/dochodok'
 
-export { validate }
 export default Dochodok
