+++ src/pages/dve-percenta.tsx
@@ -1,4 +1,3 @@
-import { validate, DvePercenta } from '../plugins/steps/3d-dve-percenta/dve-percenta'
+import { DvePercenta } from '../plugins/steps/3d-dve-percenta/dve-percenta'
 
-export { validate }
 export default DvePercenta
