+++ src/pages/hypoteka.tsx
@@ -1,4 +1,3 @@
-import { validate, Hypoteka } from '../plugins/steps/3c-hypoteka/hypoteka'
+import { Hypoteka } from '../plugins/steps/3c-hypoteka/hypoteka'
 
-export { validate }
 export default Hypoteka
