+++ src/pages/index.tsx
@@ -1,6 +1,3 @@
-
 import { Home } from '../plugins/main/home/Home'
 
 export default Home
-
-
