+++ src/pages/osobne-udaje.tsx
@@ -1,4 +1,3 @@
-import { validate, OsobneUdaje } from '../plugins/steps/5-osobne-udaje/osobne-udaje'
+import { OsobneUdaje } from '../plugins/steps/5-osobne-udaje/osobne-udaje'
 
-export { validate }
 export default OsobneUdaje
