+++ src/pages/partner.tsx
@@ -1,4 +1,3 @@
-import { validate, Partner } from '../plugins/steps/3-manzelstvo/partner'
+import { Partner } from '../plugins/steps/3-manzelstvo/partner'
 
-export { validate }
 export default Partner
