+++ src/pages/prijmy-a-vydavky.tsx
@@ -1,4 +1,3 @@
-import { validate, PrijmyAVydavky } from '../plugins/steps/1-prijmy/prijmy-a-vydavky'
+import { PrijmyAVydavky } from '../plugins/steps/1-prijmy/prijmy-a-vydavky'
 
-export { validate }
 export default PrijmyAVydavky
