+++ src/pages/zamestnanie.tsx
@@ -1,4 +1,3 @@
-import { validate, Zamestnanie } from '../plugins/steps/2-zamestnanie/zamestnanie'
+import { Zamestnanie } from '../plugins/steps/2-zamestnanie/zamestnanie'
 
-export { validate }
 export default Zamestnanie
