+++ src/plugins/_REFACTOR.md
@@ -1,0 +1,6 @@
+
+# 30.12. roman
+- porozdelovat komponenty ktore patria len k jednotlivym pages
+- calculation
+- pozriet reuse v suboroch
+- doplnit dev tlacitka na predvyplnenie pocas vyvoja
