+++ src/plugins/main/home/PostponeSection.js
@@ -1,5 +1,5 @@
 import Link from 'next/link'
-import { TAX_YEAR } from '../../../lib/calculation'
+import { TAX_YEAR } from '../../summary/calculation/calculation'
 
 export const PostponeSection = ({ nextPostponeRoute }) => (
     <>
