+++ src/plugins/steps/1-prijmy/prijmy-a-vydavky.tsx
@@ -6,7 +6,7 @@ import { numberInputRegexp } from '../../../lib/utils'
 import { ErrorSummary } from '../../_components/form/ErrorSummary'
 import { Page } from '../../../components/Page'
 import { BackLink } from '../../_components/links/BackLink'
-import { TAX_YEAR } from '../../../lib/calculation'
+import { TAX_YEAR } from '../../summary/calculation/calculation'
 
 const PrijmyAVydavky: Page<IncomeAndExpenseUserInput> = ({
   taxFormUserInput,
