+++ src/plugins/summary/6-suhrn/suhrn.module.css
